Fix stack underflow on top-level closure calls
call_closure_with computed slot_start with usize subtraction, but the
value is legitimately negative for a top-level script's own frame
(cast to isize right after). Any call made before the stack holds
another value - e.g. a plain closure invoked as the program's first
statement - underflowed the subtraction and panicked instead of
producing slot_start = -1.

Compute the subtraction in isize throughout.

+++ src/vm/functions.rs
@@ -408,7 +408,7 @@ impl VirtualMachine {
         // Stack layout: [...previous..., arg0, arg1, ..., argN, func_obj]
         // slot_start should point just BEFORE the first argument
         // So: slot_start = current_len - arg_count - 1 (for func) - 1 (to go before first arg)
-        let slot_start = (self.stack.len() - arg_count - 1 - 1) as isize;
+        let slot_start = self.stack.len() as isize - arg_count as isize - 1 - 1;
 
         let new_frame = CallFrame {
             closure: Rc::clone(closure),
